refactor(explorer): remove cardano db v1 support
