+++ mithril-explorer/src/app/page.js
@@ -17,7 +17,6 @@ import {
 import ControlPanel from "#/ControlPanel";
 import CardanoBlocksTransactionsSnapshotsList from "#/Artifacts/CardanoBlocksTransactionsSnapshotsList";
 import CardanoDbSnapshotsList from "#/Artifacts/CardanoDbSnapshotsList";
-import CardanoDbV2SnapshotsList from "#/Artifacts/CardanoDbV2SnapshotsList";
 import CardanoStakeDistributionsList from "#/Artifacts/CardanoStakeDistributionsList";
 import CardanoTransactionsSnapshotsList from "#/Artifacts/CardanoTransactionsSnapshotsList";
 import CertificatesList from "#/Artifacts/CertificatesList";
@@ -46,7 +45,7 @@ export default function Explorer() {
   const [enableCardanoTransactionTab, setEnableCardanoTransactionTab] = useState(false);
   const [enableCardanoBlocksTransactionTab, setEnableCardanoBlocksTransactionTab] = useState(false);
   const [enableCardanoStakeDistributionTab, setEnableCardanoStakeDistributionTab] = useState(false);
-  const [enableCardanoDbV2Tab, setEnableCardanoDbV2Tab] = useState(false);
+  const [enableCardanoDbTab, setEnableCardanoDbTab] = useState(false);
   const [currentTab, setCurrentTab] = useState(defaultTab);
   const selectedAggregator = useSelector(currentlySelectedAggregator);
   const selectedAggregatorSignedEntities = useSelector((state) =>
@@ -63,7 +62,7 @@ export default function Explorer() {
     setEnableCardanoStakeDistributionTab(
       selectedAggregatorSignedEntities.includes(signedEntityType.CardanoStakeDistribution),
     );
-    setEnableCardanoDbV2Tab(selectedAggregatorSignedEntities.includes(signedEntityType.CardanoDb));
+    setEnableCardanoDbTab(selectedAggregatorSignedEntities.includes(signedEntityType.CardanoDb));
   }, [selectedAggregatorSignedEntities]);
 
   useEffect(() => {
@@ -116,12 +115,9 @@ export default function Explorer() {
     <Stack gap={3}>
       <ControlPanel />
       <Tabs activeKey={currentTab} onSelect={(key) => setCurrentTab(key)}>
-        <Tab title="Cardano Db" eventKey={signedEntityType.CardanoImmutableFilesFull}>
-          <CardanoDbSnapshotsList />
-        </Tab>
-        {enableCardanoDbV2Tab && (
-          <Tab title="Cardano Db v2" eventKey={signedEntityType.CardanoDb}>
-            <CardanoDbV2SnapshotsList />
+        {enableCardanoDbTab && (
+          <Tab title="Cardano Db" eventKey={signedEntityType.CardanoDb}>
+            <CardanoDbSnapshotsList />
           </Tab>
         )}
         {enableCardanoTransactionTab && (
