+++ mithril-explorer/src/constants.js
@@ -3,7 +3,6 @@ export const aggregatorSearchParam = "aggregator";
 export const signedEntityType = {
   MithrilStakeDistribution: "MithrilStakeDistribution",
   CardanoStakeDistribution: "CardanoStakeDistribution",
-  CardanoImmutableFilesFull: "CardanoImmutableFilesFull",
   CardanoDb: "CardanoDatabase",
   CardanoTransactions: "CardanoTransactions",
   CardanoBlocksTransactions: "CardanoBlocksTransactions",
